+++ NEWS.md
@@ -4,7 +4,8 @@
 # text 0.9.17
 ### New Features
 - textrpp_initiate() runs automatically in library(text) when default environment exits
-- option to capture python warnings
+- Python warnings a captured in embedding comments
+- Option to print python options to console